Refactors IP exclusion logic and enhances uptime monitoring
This commit refines the IP exclusion functionality by moving the `isIPExcluded` method to the `siteConfig` class, improving code organization and reusability. Additionally, it updates the uptime monitoring service initialization to ensure it runs correctly in the background, enhancing the application's reliability.

Changes include:
- Moved IP exclusion check to `siteConfig` for better encapsulation.
- Updated the uptime monitoring service to remove commented-out code and ensure proper logging on initialization success or failure.
