+++ server/src/api/sites/updateSiteExcludedIPs.ts
@@ -35,17 +35,9 @@ export async function updateSiteExcludedIPs(request: FastifyRequest, reply: Fast
       });
     }
 
-    // Check if user is authenticated
-    if (!request.user?.id) {
-      return reply.status(401).send({
-        success: false,
-        error: "Authentication required",
-      });
-    }
-
     // Check if user has access to this site
     const hasAccess = await getUserHasAccessToSite(request, numericSiteId);
-    
+
     if (!hasAccess) {
       return reply.status(403).send({
         success: false,
